pyproject.toml
[project]
name = "simplestockdemo"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "django>=5.2.1",
    "djangorestframework>=3.16.0",
    "hiredis>=3.1.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "redis>=6.1.0",
]
